code review comments lead to spaceport re-structure
Restructured the code as follows:

 - spaceport module moved to server module in own file (server.rs)
 - delete the redundant client as no longer providing any value
 - use SocketAddr directly in clap Args
